feat: 다크 모드 구현 홈 화면에서는 footer만 다크모드 적용 나머지 화면 light, dark 모드 구현 완료 색상은 추가로 설정 필요 src 디렉토리 생성, 정리함.
