+++ .env
@@ -1,0 +1,1 @@
+RECOIL_DUPLICATE_ATOM_KEY_CHECKING_ENABLED=false
