+++ package.json
@@ -21,6 +21,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.8.0",
+    "recoil": "^0.7.7",
     "styled-components": "^5.3.8",
     "typescript": "4.9.5"
   }
