--- pages/album.tsx
@@ -1,13 +1,0 @@
-import Header from "@/components/Header";
-import PageTitle from "@/components/PageTitle";
-
-function Album() {
-  return (
-    <>
-      <Header bgColor="transparent" />
-      <PageTitle pageTitle="Album"></PageTitle>
-    </>
-  );
-}
-
-export default Album;
