--- pages/boards.tsx
@@ -1,14 +1,0 @@
-import Header from "@/components/Header";
-import PageTitle from "@/components/PageTitle";
-
-function Boards() {
-  return (
-    <>
-      <Header bgColor="transparent" />
-
-      <PageTitle pageTitle="Boards"></PageTitle>
-    </>
-  );
-}
-
-export default Boards;
