+++ src/atoms.ts
@@ -1,0 +1,6 @@
+import { atom } from "recoil";
+
+export const isDarkState = atom({
+  key: "isDark",
+  default: true,
+});
