+++ src/components/PageTitle.tsx
@@ -3,17 +3,15 @@ import styled from "styled-components";
 const Wrapper = styled.div`
   height: 120px;
   width: 100vw;
-  border-top: 1px solid white;
-  border-bottom: 1px solid white;
+  border-top: 1px solid ${(props) => props.theme.textColor};
+  border-bottom: 1px solid ${(props) => props.theme.textColor};
   box-sizing: border-box;
-  margin-top: 71px;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
   gap: 1em;
 `;
-
 const Title = styled.div`
   font-size: 4.8rem;
 `;
@@ -25,6 +23,7 @@ const Sub = styled.div`
 interface pageTitleProps {
   pageTitle: string;
 }
+
 function PageTitle({ pageTitle }: pageTitleProps) {
   return (
     <Wrapper>
