+++ src/components/SideBar.tsx
@@ -1,4 +1,5 @@
 import styled from "styled-components";
+
 const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
