+++ src/pages/album.tsx
@@ -1,0 +1,11 @@
+import PageTitle from "../components/PageTitle";
+
+function Album() {
+  return (
+    <>
+      <PageTitle pageTitle="Album"></PageTitle>
+    </>
+  );
+}
+
+export default Album;
