+++ src/pages/boards.tsx
@@ -1,0 +1,11 @@
+import PageTitle from "../components/PageTitle";
+
+function Boards() {
+  return (
+    <>
+      <PageTitle pageTitle="Boards"></PageTitle>
+    </>
+  );
+}
+
+export default Boards;
