+++ src/styles/styled.d.ts
