+++ src/styles/theme.ts
@@ -6,3 +6,10 @@ export const darkTheme: DefaultTheme = {
   color2: "#485460", // 연함
   textColor: "#FFFFFF",
 };
+
+export const lightTheme: DefaultTheme = {
+  bgColor: "#d9e1e8",
+  color1: "#12171C", // 진함
+  color2: "#9baec8", // 연함
+  textColor: "#282c37",
+};
